+++ src/Simple.Migrations/AssemblyMigrationProvider.cs
@@ -2,27 +2,30 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
-using SimpleMigrations.Platform;
 
 namespace SimpleMigrations
 {
     /// <summary>
-    /// <see cref="IMigrationProvider"/>  which finds migrations by scanning an assembly
+    /// <see cref="IMigrationProvider"/>  which finds migrations by scanning an assembly, optionally
+    /// filtering by namespace.
     /// </summary>
     public class AssemblyMigrationProvider : IMigrationProvider
     {
         private readonly Assembly migrationAssembly;
+        private readonly string migrationNamespace;
 
         /// <summary>
         /// Instantiates a new instance of the <see cref="AssemblyMigrationProvider"/> class
         /// </summary>
         /// <param name="migrationAssembly">Assembly to scan for migrations</param>
-        public AssemblyMigrationProvider(Assembly migrationAssembly)
+        /// <param name="migrationNamespace">Optional namespace. If specified, only finds migrations in that namespace</param>
+        public AssemblyMigrationProvider(Assembly migrationAssembly, string migrationNamespace = null)
         {
             if (migrationAssembly == null)
                 throw new ArgumentNullException(nameof(migrationAssembly));
 
             this.migrationAssembly = migrationAssembly;
+            this.migrationNamespace = migrationNamespace;
         }
 
         /// <summary>
@@ -31,10 +34,11 @@ namespace SimpleMigrations
         /// <returns>All migration info</returns>
         public IEnumerable<MigrationData> LoadMigrations()
         {
-            var migrations = from type in this.migrationAssembly.GetDefinedTypes()
+            var migrations = from type in this.migrationAssembly.DefinedTypes
                              let attribute = type.GetCustomAttribute<MigrationAttribute>()
                              where attribute != null
-                             select new MigrationData(attribute.Version, attribute.Description, type.AsType());
+                             where this.migrationNamespace == null || type.Namespace == this.migrationNamespace
+                             select new MigrationData(attribute.Version, attribute.Description, type);
 
             if (!migrations.Any())
                 throw new MigrationException($"Could not find any migrations in the assembly you provided ({this.migrationAssembly.GetName().Name}). Migrations must be decorated with [Migration]");