+++ src/Simple.Migrations/DatabaseProvider/MssqlDatabaseProvider.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Data;
 using System.Data.Common;
 
 namespace SimpleMigrations.DatabaseProvider