+++ src/Simple.Migrations/ISimpleMigrator.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections.Generic;
-using System.Collections.ObjectModel;
 
 namespace SimpleMigrations
 {
@@ -26,11 +25,7 @@ namespace SimpleMigrations
         /// <summary>
         /// Gets all available migrations
         /// </summary>
-#if NET40
-        ReadOnlyCollection<MigrationData> Migrations { get; }
-#else
         IReadOnlyList<MigrationData> Migrations { get; }
-#endif
 
         /// <summary>
         /// Load all available migrations, and the current state of the database